Enable differential loading
This also removes the need for systemjs in modern browsers.
Now we can use dynamic import to load our routes, and there is no longer a need for different app module for dev and prod.

For older browsers, we serve core-js and systemjs to polyfill the loader
and ES2015 APIs.