Make route changes obvious
In the demo prior to the change, route changes (i.e. when a user clicks
a nav link) appear to be broken: nothing changes on the screen. The
success is visible only while looking at the browser JS console.

This change very slightly updates the components to make the topmost
heading on the main pane change to match the navigation click.

(Discovered the value of this during internal Bazel+Angular
walkthroughs.)